+++ dependencies/ominipg.ts
@@ -1,1 +1,1 @@
-export * from "jsr:@oxian/ominipg@0.9.0-rc.9";
+export * from "jsr:@oxian/ominipg@0.9.0-rc.10";