+++ dependencies/oxian-work.ts
@@ -1,1 +1,1 @@
-export * from "jsr:@oxian/oxian-js@0.21.0-rc.3/work";
+export * from "jsr:@oxian/oxian-js@0.21.0-rc.4/work";